+++ sphinx/screens/chats/chat-tribe/chat-tribe/src/main/java/chat/sphinx/chat_tribe/ui/ChatTribeFragment.kt
@@ -14,7 +14,6 @@ import chat.sphinx.chat_common.ui.ChatFragment
 import chat.sphinx.chat_common.ui.viewstate.messagereply.MessageReplyViewState
 import chat.sphinx.chat_tribe.R
 import chat.sphinx.chat_tribe.databinding.FragmentChatTribeBinding
-import chat.sphinx.chat_tribe.databinding.LayoutPodcastPlayerFooterBinding
 import chat.sphinx.chat_tribe.model.TribeFeedData
 import chat.sphinx.concept_image_loader.ImageLoader
 import chat.sphinx.concept_image_loader.ImageLoaderOptions
@@ -22,12 +21,14 @@ import chat.sphinx.concept_image_loader.Transformation
 import chat.sphinx.concept_user_colors_helper.UserColorsHelper
 import chat.sphinx.menu_bottom.databinding.LayoutMenuBottomBinding
 import chat.sphinx.resources.databinding.LayoutBoostFireworksBinding
+import chat.sphinx.resources.databinding.LayoutPodcastPlayerFooterBinding
 import chat.sphinx.resources.getString
 import chat.sphinx.wrapper_common.lightning.asFormattedString
 import chat.sphinx.wrapper_message.*
 import chat.sphinx.wrapper_view.Px
 import dagger.hilt.android.AndroidEntryPoint
 import io.matthewnelson.android_feature_screens.util.gone
+import io.matthewnelson.android_feature_screens.util.goneIfFalse
 import io.matthewnelson.android_feature_screens.util.visible
 import io.matthewnelson.concept_views.viewstate.collect
 import io.matthewnelson.concept_views.viewstate.value
@@ -109,22 +110,25 @@ internal class ChatTribeFragment: ChatFragment<
         }
 
         podcastPlayerBinding.apply {
-            textViewBoostPodcastButton.setOnClickListener {
-                tribeFeedViewModel.podcastViewStateContainer.value.clickBoost?.let {
-                    it.invoke()
-                    boostAnimationBinding.apply {
-                        root.visible
-                        lottieAnimationView.playAnimation()
-                    }
-                }
-            }
+//            textViewBoostPodcastButton.setOnClickListener {
+//                tribeFeedViewModel.podcastViewStateContainer.value.clickBoost?.let {
+//                    it.invoke()
+//                    boostAnimationBinding.apply {
+//                        root.visible
+//                        lottieAnimationView.playAnimation()
+//                    }
+//                }
+//            }
             textViewForward30Button.setOnClickListener {
                 tribeFeedViewModel.podcastViewStateContainer.value.clickFastForward?.invoke()
             }
-            textViewPlayPauseButton.setOnClickListener {
+            textViewPlayButton.setOnClickListener {
                 tribeFeedViewModel.podcastViewStateContainer.value.clickPlayPause?.invoke()
             }
-            textViewEpisodeTitle.setOnClickListener {
+            animationViewPauseButton.setOnClickListener {
+                tribeFeedViewModel.podcastViewStateContainer.value.clickPlayPause?.invoke()
+            }
+            layoutConstraintPodcastInfo.setOnClickListener {
                 tribeFeedViewModel.podcastViewStateContainer.value.clickTitle?.invoke()
             }
         }
@@ -198,25 +202,30 @@ internal class ChatTribeFragment: ChatFragment<
                             root.gone
                         }
                         is PodcastViewState.PodcastVS -> {
-                            textViewPlayPauseButton.text = if (viewState.showPlayButton) {
-                                getString(R.string.material_icon_name_play_button)
-                            } else {
-                                getString(R.string.material_icon_name_pause_button)
-                            }
+
+                            textViewPlayButton.goneIfFalse(viewState.showPlayButton && !viewState.showLoading)
+                            animationViewPauseButton.goneIfFalse(!viewState.showPlayButton && !viewState.showLoading)
 
                             val calculatedWidth = progressWidth.value.toDouble() * (viewState.playingProgress / 100.0)
                             progressBar.layoutParams.width = calculatedWidth.toInt()
                             progressBar.requestLayout()
 
+                            textViewEpisodeTitle.isSelected = !viewState.showPlayButton && !viewState.showLoading
                             textViewEpisodeTitle.text = viewState.title
+                            textViewContributorTitle.text = viewState.subtitle
 
-                            if (viewState.showLoading) {
-                                textViewPlayPauseButton.isEnabled = false
-                                progressBarAudioLoading.visible
-                            } else {
-                                textViewPlayPauseButton.isEnabled = true
-                                progressBarAudioLoading.gone
+                            viewState.imageUrl?.let { imageUrl ->
+                                imageLoader.load(
+                                    imageViewPodcastEpisode,
+                                    imageUrl,
+                                    ImageLoaderOptions.Builder()
+                                        .placeholderResId(chat.sphinx.podcast_player.R.drawable.ic_profile_avatar_circle)
+                                        .build()
+                                )
                             }
+
+                            textViewForward30Button.goneIfFalse(!viewState.showLoading)
+                            progressBarAudioLoading.goneIfFalse(viewState.showLoading)
 
                             scrollToBottom(callback = {
                                 root.visible