+++ sphinx/screens/dashboard/dashboard/src/main/java/chat/sphinx/dashboard/ui/adapter/DashboardFooterAdapter.kt
@@ -1,0 +1,35 @@
+package chat.sphinx.dashboard.ui.adapter
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import chat.sphinx.dashboard.databinding.LayoutDashboardFooterBinding
+import chat.sphinx.insetter_activity.InsetterActivity
+import chat.sphinx.insetter_activity.addNavigationBarPadding
+
+/**
+ * Needed in order to have the last item of the RecyclerView
+ * be able to scroll up over the device's navigation bar.
+ * */
+internal class DashboardFooterAdapter: RecyclerView.Adapter<DashboardFooterAdapter.FooterViewHolder>() {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FooterViewHolder {
+        val binding = LayoutDashboardFooterBinding.inflate(
+            LayoutInflater.from(parent.context),
+            parent,
+            false
+        )
+
+        return FooterViewHolder(binding)
+    }
+
+    override fun onBindViewHolder(holder: FooterViewHolder, position: Int) {}
+
+    override fun getItemCount(): Int {
+        return 1
+    }
+
+    inner class FooterViewHolder(
+        binding: LayoutDashboardFooterBinding
+    ): RecyclerView.ViewHolder(binding.root)
+}