+++ sphinx/screens/dashboard/dashboard/src/main/java/chat/sphinx/dashboard/ui/feed/read/FeedReadFragment.kt
@@ -1,5 +1,6 @@
 package chat.sphinx.dashboard.ui.feed.read
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.os.Bundle
 import android.view.View
@@ -11,6 +12,7 @@ import chat.sphinx.dashboard.R
 import chat.sphinx.dashboard.databinding.FragmentFeedReadBinding
 import chat.sphinx.dashboard.ui.adapter.FeedFollowingAdapter
 import chat.sphinx.dashboard.ui.adapter.FeedReadNowAdapter
+import chat.sphinx.dashboard.ui.feed.FeedFragment
 import chat.sphinx.dashboard.ui.viewstates.FeedReadViewState
 import dagger.hilt.android.AndroidEntryPoint
 import io.matthewnelson.android_feature_screens.ui.sideeffect.SideEffectFragment
@@ -38,8 +40,19 @@ internal class FeedReadFragment : SideEffectFragment<
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        setupNestedScrollView()
         setupReadNowAdapter()
         setupFollowingAdapter()
+    }
+
+    @SuppressLint("RestrictedApi")
+    private fun setupNestedScrollView() {
+        binding.scrollViewContent.setOnScrollChangeListener { _, _, scrollY, _, oldScrollY ->
+            if (parentFragment is FeedFragment) {
+                val offsetY = binding.scrollViewContent.computeVerticalScrollOffset()
+                (parentFragment as FeedFragment)?.shouldToggleNavBar(scrollY < oldScrollY && offsetY < 50)
+            }
+        }
     }
 
     private fun setupReadNowAdapter() {